+++ intervalues/abstract_interval.py
@@ -1,75 +1,112 @@
 import abc
-import random
 import intervalues
+from typing import Iterator, Optional, Counter, TypeVar
+
+T = TypeVar('T', bound='intervalues.BaseInterval')
 
 
 class AbstractInterval(abc.ABC):
-    pass
+    """
+    Abstract class for intervals of any type: a single base interval, or a collection of intervals in some way.
 
-
-# TODO: implement Set -> move same implementations for Set and Counter to below -> make sure shared between B/S/C is in
-# AI above.
-
-class AbstractIntervalCollector(AbstractInterval):
+    Contains self-explaining methods for:
+    - converting the object to a IntervalCounter/IntervalList/IntervalMeter
+    - calculating some general interval properties, the max/min and the length/weight of it
+    """
 
     @abc.abstractmethod
-    def __init__(self, data=None):
-        self.data = (None,) if data is None else data
+    def as_counter(self) -> 'intervalues.IntervalCounter': pass
 
-    def get_data(self):
+    @abc.abstractmethod
+    def as_list(self) -> 'intervalues.IntervalList': pass
+
+    @abc.abstractmethod
+    def as_meter(self) -> 'intervalues.IntervalMeter': pass
+
+    @abc.abstractmethod
+    def as_set(self) -> 'intervalues.IntervalSet': pass
+
+    @abc.abstractmethod
+    def as_pdf(self) -> 'intervalues.IntervalPdf': pass
+
+    @abc.abstractmethod
+    def get_length(self) -> float: pass
+
+    @abc.abstractmethod
+    def max(self) -> float: pass
+
+    @abc.abstractmethod
+    def min(self) -> float: pass
+
+
+class AbstractIntervalCollection(AbstractInterval):
+    """
+    Abstract class for interval collections of intervals in some way.
+    In general, the relevant data for each collection wil be contained in a `data` attribute.
+
+    Contains methods for:
+    - accessing/defining/changing the contents of `data`
+    - comparing with other objects
+    - converting to a base interval
+    """
+
+    @abc.abstractmethod
+    def __init__(self, data: Optional[Counter | list | set] = None):
+        self.data: Counter | list | set = list() if data is None else data
+
+    def get_data(self) -> Counter | list | set:
         return self.data
 
-    def set_data(self, data):
+    def set_data(self, data: Counter | list | set):
         self.data = data
 
-    def get_length(self):
-        return len(self.data)
+    @abc.abstractmethod
+    def get_length(self) -> float:
+        pass
 
-    def __len__(self):
-        return len(self.data)
-
-    def sample(self, k=1):
-        return random.sample(self.data, k=k)
-
-    def draw(self, k=1):
-        return self.sample(k=k)
-
-    def __contains__(self, x):
+    def __contains__(self, x: 'intervalues.BaseInterval | float') -> bool:
         return x in self.data
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return f"{self.__class__}:{self.data}"
 
-    def __str__(self):
+    def __str__(self) -> str:
         return self.__repr__()
 
-    def __getitem__(self, x):
-        return self.data[x]
+    @abc.abstractmethod
+    def __getitem__(self, x) -> float:
+        pass
 
-    def __eq__(self, other):
+    def __eq__(self, other: object) -> bool:
         return isinstance(other, type(self)) and (self.data == other.data)
 
-    def __hash__(self):
+    def __hash__(self) -> int:
         return hash(tuple(self))
 
-    def __iter__(self):
+    def __iter__(self) -> Iterator:
         return iter(self.data)
 
-    def __add__(self, other):
-        return __class__(self.data + other.data)
+    @abc.abstractmethod
+    def __add__(self, other: 'T | AbstractIntervalCollection') -> \
+            'T | AbstractIntervalCollection':
+        pass
 
-    def __iadd__(self, other):
-        self.data += other.data
+    @abc.abstractmethod
+    def __mul__(self, other: float) -> 'AbstractIntervalCollection':
+        pass
 
+    def __neg__(self) -> 'AbstractIntervalCollection':
+        return self.__mul__(-1)
+
+    @abc.abstractmethod
     def update(self, data):
-        self.data += data
+        pass
 
-    def min(self):
-        return min(self.data)
+    def min(self) -> float:
+        return min(min(x) for x in self.data)
 
-    def max(self):
-        return max(self.data)
+    def max(self) -> float:
+        return max(max(x) for x in self.data)
 
-    # @abc.abstractmethod
-    def as_single_interval(self):
+    def as_single_interval(self) -> 'intervalues.BaseInterval':
         return intervalues.BaseInterval(self.min(), self.max())