+++ intervalues/interval_pdf.py
@@ -1,10 +1,29 @@
+from typing import Optional, Sequence
+
 import intervalues
 from random import random
 
 
 class IntervalPdf(intervalues.IntervalMeter):
+    __name__ = 'IntervalPdf'
 
-    def __init__(self, data=None):
+    """
+    Class for a probability density function across intervals, that can be used for sampling and other statistical 
+    purposes.
+
+    Objects can be instantiated in multiple ways (with `a = BaseInterval((1, 3))` and `b = BaseInterval((0, 2))`):
+    - IntervalPdf(a) -> using a single interval
+    - IntervalPdf([a, b]) -> using a list, tuple or set of intervals
+
+    The data is collected in a standard Counter. For the keys, the BaseIntervals are converted to value=1, and the value
+    is tracked using the value of the Counter. In contrast to IntervalMeters, the values of IntervalPdfs will 
+    automatically be scaled such that the total length equals 1, like a probability density should.
+
+    All methods available for Counters (most_common, items, etc) are available, as well as all IntervalCollection
+    methods (get_length, max, etc). Use .cumulative to convert a IntervalPdf into a Cdf, or use sample to draw a random
+    value from any subinterval in the IntervalPdf using the normalized value as density.
+    """
+    def __init__(self, data: Optional[Sequence['intervalues.BaseInterval'] | 'intervalues.BaseInterval'] = None):
         super().__init__(data)
         self.normalize()
 
@@ -13,29 +32,29 @@ class IntervalPdf(intervalues.IntervalMeter):
         for k, v in self.items():
             self.data[k] = v / total
 
-    def pop(self, __key):
+    def pop(self, __key: 'intervalues.BaseInterval') -> float:
         item = self.data.pop(__key)
         self.normalize()
         return item
 
-    def popitem(self):
+    def popitem(self) -> tuple['intervalues.BaseInterval', float]:
         item = self.data.popitem()
         self.normalize()
         return item
 
-    def total_length(self, force=False):
+    def total_length(self, force: bool = False) -> float:
         if not force:
             return 1
         return super().total_length()
 
-    def __mul__(self, other):
+    def __mul__(self, other: float) -> 'IntervalPdf':
         return self.copy()
 
-    def __imul__(self, other):
+    def __imul__(self, other: float) -> 'IntervalPdf':
         return self
 
-    def __repr__(self):
-        return f"IntervalPDF:{dict(self.data)}"
+    def __repr__(self) -> str:
+        return f"{self.__name__}:{dict(self.data)}"
 
     def check_intervals(self):
         super().check_intervals()
@@ -46,35 +65,54 @@ class IntervalPdf(intervalues.IntervalMeter):
         super().align_intervals()
         self.normalize()
 
-    def cumulative(self, x):
+    def cumulative(self, x: float) -> float:
         pre = sum([self.get_length(i) for i in self.keys() if i.max() < x])
-        this = self.find_which_contains(x)
-        if this:
-            this = self.get_length(this) * (x - this.min()) / (this.max() - this.min())
-        return pre + this
+        this: 'bool | intervalues.BaseInterval' = self.find_which_contains(x)
+        if isinstance(this, intervalues.BaseInterval):
+            this_val = self.get_length(this) * (x - this.min()) / (this.max() - this.min())
+        else:
+            this_val = 0
+        return pre + this_val
 
-    def cumsum(self, x):
+    def cumsum(self, x: float) -> float:
         return self.cumulative(x)
 
-    def inverse_cumulative(self, p):
+    def inverse_cumulative(self, p: float) -> float:
+        # Note: here the inverse-CDF sampling method is used. Alternatively, we could a combination of random.choice to
+        # select a subinterval and then random() to sample within that subinterval in an uniform way.
+
         keys = sorted(self.keys())
-        sum_p, iter, last = 0, -1, 0
+        i: int = -1
+        sum_p: float = 0
+        last: float = 0
         while sum_p < p:
-            iter += 1
+            i += 1
             last = sum_p
-            sum_p += self.get_length(keys[iter])
-        if iter == -1:
+            sum_p += self.get_length(keys[i])
+        if i == -1:
             return 0
 
-        where_in_curr = (p - last) / self.get_length(keys[iter])
-        min_curr, max_curr = keys[iter].min(), keys[iter].max()
+        where_in_curr = (p - last) / self.get_length(keys[i])
+        min_curr, max_curr = keys[i].min(), keys[i].max()
         x = where_in_curr * (max_curr - min_curr) + min_curr
-        print(where_in_curr, keys[iter], min_curr, max_curr)
+        print(where_in_curr, keys[i], min_curr, max_curr)
 
         return x
 
-    def sample(self, k=1):
-        return (self.inverse_cumulative(random()) for _ in range(k))
+    def sample(self, k: int = 1) -> list[float]:
+        return [self.inverse_cumulative(random()) for _ in range(k)]
 
-    def as_meter(self):
+    def as_meter(self) -> 'intervalues.IntervalMeter':
         return intervalues.IntervalMeter(tuple(self))
+
+    @staticmethod
+    def as_my_type(other: 'intervalues.AbstractInterval') -> 'IntervalPdf':
+        return other.as_pdf()
+
+    def copy(self) -> 'IntervalPdf':
+        return self.__copy__()
+
+    def __copy__(self) -> 'IntervalPdf':
+        new_counter = self.__class__()
+        new_counter.data = self.data.copy()
+        return new_counter