+++ intervalues/interval_set.py
@@ -1,126 +1,151 @@
-import intervalues
+import collections
+from typing import Optional, Sequence, Iterator
+
 from intervalues import base_interval
-from intervalues.abstract_interval import AbstractIntervalCollector
-from intervalues.combine_intervals import combine_intervals_set, combine_intervals_meter
+from intervalues.abstract_interval import AbstractIntervalCollection
+from intervalues.combine_intervals import combine_intervals_set, combine_intervals_meter, combine_intervals_set_discrete
+import intervalues
 
 
-class IntervalSet(AbstractIntervalCollector):
+class IntervalSet(AbstractIntervalCollection):
+    __name__ = 'IntervalSet'
 
-    def __init__(self, data=None):
+    """
+    Class for a set of intervals, that tracks the occurence of individual subintervals across its inputs.
+
+    Objects can be instantiated in multiple ways (with `a = BaseInterval((1, 3))` and `b = BaseInterval((0, 2))`):
+    - IntervalSet(a) -> using a single interval
+    - IntervalSet([a, b]) -> using a list, tuple or set of intervals
+
+    The data is collected in a standard set. For this, the BaseIntervals are converted to value=1, since the IntervalSet
+    doesn't track how often subintervals are featured.
+
+    All methods available for sets (intersection, symmetric_difference, etc) are available, as well as all 
+    IntervalCollection methods (get_length, max, etc). IntervalSets can be combined together (union/intersection) or 
+    differenced. They can also be converted to IntervalCounters, IntervalMeters or IntervalLists. 
+    """
+
+    def __init__(self, data: Optional[Sequence['intervalues.BaseInterval'] | 'intervalues.BaseInterval'] = None):
         super().__init__()
-        self.data = set()
+        self.data: set = set()
         if data is not None:
-            if type(data) in (list, tuple, set):
-                combine_intervals_set(data, object_exists=self)
-            elif type(data) is base_interval.BaseInterval:
-                self.data = {data.as_index()}
+            self.discrete: bool = True if type(data) == Sequence[
+                intervalues.BaseDiscreteInterval] | intervalues.BaseDiscreteInterval else False
+            if self.discrete:
+                if isinstance(data, collections.abc.Sequence):
+                    combine_intervals_set_discrete(data, object_exists=self)  # type: ignore[arg-type]
+                elif type(data) is intervalues.BaseDiscreteInterval:
+                    self.data = {data.as_index()}
             else:
-                combine_intervals_set(tuple(data), object_exists=self)
+                if isinstance(data, collections.abc.Sequence):
+                    combine_intervals_set(data, object_exists=self)
+                elif type(data) is base_interval.BaseInterval:
+                    self.data = {data.as_index()}
 
-    def add(self, other):  # reliably restored by inspect
-        self.update_interval(other)
+    def add(self, other: 'IntervalSet'):
+        self.update_set(other)
 
-    def difference(self, other):  # reliably restored by inspect
+    def difference(self, other: 'IntervalSet') -> 'IntervalSet':
         return self - other
 
-    def difference_update(self, other):  # reliably restored by inspect
+    def difference_update(self, other: 'IntervalSet'):
         self.__isub__(other)
 
-    def discard(self, item):  # reliably restored by inspect
-        self.data -= item
+    def discard(self, item: 'IntervalSet'):
+        self.data -= item.data
 
-    def intersection(self, other):  # reliably restored by inspect
+    def intersection(self, other: 'IntervalSet') -> 'IntervalSet':
         new = self.copy()
         new += other
         return new
 
-    def intersection_update(self, other):  # reliably restored by inspect
+    def intersection_update(self, other: 'IntervalSet'):
         self.__iadd__(other)
 
-    def isdisjoint(self, other):  # reliably restored by inspect
+    def isdisjoint(self, other: 'IntervalSet') -> bool:
         return all([x.is_disjoint_with(y) for x in self.data for y in other.data])
 
-    def issubset(self, other):  # reliably restored by inspect
+    def issubset(self, other: 'IntervalSet') -> bool:
         return all([any([x in y for y in other.data]) for x in self.data])
 
-    def issuperset(self, other):  # reliably restored by inspect
+    def issuperset(self, other: 'IntervalSet') -> bool:
         return other.issubset(self)
 
-    def pop(self):  # reliably restored by inspect
+    def pop(self) -> 'intervalues.BaseInterval':
         return self.data.pop()
 
-    def remove(self, item):  # reliably restored by inspect
+    def remove(self, item: 'intervalues.BaseInterval'):
         if item not in self.data:
             raise KeyError(f"{item} not in {self}")
         self.data.remove(item)
 
-    def symmetric_difference(self, other):  # reliably restored by inspect
+    def symmetric_difference(self, other: 'IntervalSet') -> 'IntervalSet':
         return self ^ other
 
-    def symmetric_difference_update(self, other):  # reliably restored by inspect
+    def symmetric_difference_update(self, other: 'IntervalSet'):
         new = self.symmetric_difference(other)
         self.data = new.data
 
-    def union(self, other):  # reliably restored by inspect
+    def union(self, other: 'IntervalSet') -> 'IntervalSet':
         return self + other
 
-    def __and__(self, other):  # reliably restored by inspect
-        new = __class__()
+    def __and__(self, other: 'IntervalSet') -> 'IntervalSet':
+        new = self.__class__()
         new.data = self.data & other.data if isinstance(other, self.__class__) else \
             (other.data if other in self.data or any([other in x for x in self.data]) else set())
         return new
 
-    def __iand__(self, other):  # reliably restored by inspect
+    def __iand__(self, other: 'IntervalSet') -> 'IntervalSet':
         if isinstance(other, self.__class__):
             self.data &= other.data
         else:
             self.data = other.data if other in self.data else set()
         return self
 
-    def __ior__(self, other):  # reliably restored by inspect
+    def __ior__(self, other: 'IntervalSet') -> 'IntervalSet':
         self.__iadd__(other)
         return self
 
-    def __ixor__(self, other):  # reliably restored by inspect
+    def __ixor__(self, other: 'IntervalSet') -> 'IntervalSet':
         new = self ^ other
         self.data = new.data
         return self
 
-    def __ne__(self, other):  # reliably restored by inspect
+    def __ne__(self, other: object) -> bool:
         return not self.__eq__(other)
 
-    def __or__(self, other):  # reliably restored by inspect
+    def __or__(self, other: 'IntervalSet') -> 'IntervalSet':
         return self + other
 
-    def __xor__(self, other):  # reliably restored by inspect
+    def __xor__(self, other: 'IntervalSet') -> 'IntervalSet':
         return (self - other) + (other - self)
 
     def clear(self):
         self.data.clear()
 
-    def copy(self):
+    def copy(self) -> 'IntervalSet':
         return self.__copy__()
 
-    def __copy__(self):
-        new_counter = __class__()
+    def __copy__(self) -> 'IntervalSet':
+        new_counter = self.__class__()
         new_counter.data = self.data.copy()
         return new_counter
 
-    def subtract(self, other):
+    def subtract(self, other: 'IntervalSet'):
         self.__isub__(other)
 
-    def total_length(self):
+    def total_length(self) -> float:
         return sum([k.get_length() for k in self.data])
 
-    def get_length(self, index=None):
+    def get_length(self, index: Optional['intervalues.BaseInterval'] = None) -> float:
         if index is None:
             return self.total_length()
         return (index in self) * index.get_length()
 
-    def __len__(self):
+    def __len__(self) -> int:
         return len(self.data)
 
-    def update(self, other, reverse=False):
+    def update(self, other: 'IntervalSet | intervalues.BaseInterval', reverse: bool = False):
         if self == other:
             if reverse:
                 self.clear()
@@ -133,11 +158,11 @@ class IntervalSet(AbstractIntervalCollector):
             raise ValueError(f'Input {other} is not of type {IntervalSet} or {base_interval.BaseInterval}')
         self.check_intervals()
 
-    def update_set(self, other, one_by_one=False, reverse=False):
+    def update_set(self, other: 'IntervalSet', one_by_one: bool = False, reverse: bool = False):
         if self == other:
             return
         else:
-            if not one_by_one:  # Join counters in one go - better for large counters with much overlap
+            if not one_by_one:  # Join sets in one go - better for large sets with much overlap
                 if not reverse:
                     combined = combine_intervals_set(list(self.data) + list(other.data))
                     self.data = combined.data
@@ -148,7 +173,7 @@ class IntervalSet(AbstractIntervalCollector):
                 for k in other.data:
                     self.update_interval(k, reverse=reverse)
 
-    def update_interval(self, other, reverse=False):
+    def update_interval(self, other: 'intervalues.BaseInterval', reverse: bool = False):
         if all([x.is_disjoint_with(other) for x in self.data]):
             if not reverse:
                 self.data.add(other)
@@ -160,7 +185,7 @@ class IntervalSet(AbstractIntervalCollector):
             if not reverse:
                 self.data.add(other)
             else:
-                combined = combine_intervals_meter(list(self.data) + [-1 * other])
+                combined = combine_intervals_set(list(self.data) + [-1 * other])
                 self.data = combined.data
             self.check_intervals()
 
@@ -177,39 +202,40 @@ class IntervalSet(AbstractIntervalCollector):
         aligned = combine_intervals_set(self_as_base)
         self.data = aligned.data
 
-    def find_which_contains(self, other):
+    def find_which_contains(self, other: 'intervalues.BaseInterval | float') -> 'bool | intervalues.BaseInterval':
         for key in self.data:
             if other in key:
                 return key
         return False
 
-    def __add__(self, other):
+    def __add__(self, other: 'intervalues.BaseInterval | AbstractIntervalCollection') -> 'IntervalSet':
         new = self.copy()
-        new.update(other)
+        new.update(other.as_set() if not isinstance(other, IntervalSet) else other)
         return new
 
-    def __iadd__(self, other):
-        self.update(other)
+    def __iadd__(self, other: 'intervalues.BaseInterval | AbstractIntervalCollection') -> 'IntervalSet':
+        self.update(other.as_set() if not isinstance(other, IntervalSet) else other)
         return self
 
-    def __sub__(self, other):
+    def __sub__(self, other: 'IntervalSet | intervalues.BaseInterval') -> 'IntervalSet':
         new = self.copy()
         new.update(other, reverse=True)
-        # new.data -= other.data
         return new
 
-    def __isub__(self, other):
+    def __isub__(self, other: 'IntervalSet | intervalues.BaseInterval') -> 'IntervalSet':
         self.update(other, reverse=True)
-        # self.data -= other.data
         return self
 
-    def __repr__(self):
-        return f"IntervalSetFloat:{self.data}"
+    def __mul__(self, other: int | float) -> 'IntervalSet':
+        return self.copy()
 
-    def __str__(self):
+    def __repr__(self) -> str:
+        return f"{self.__name__}:{self.data}"
+
+    def __str__(self) -> str:
         return self.__repr__()
 
-    def __contains__(self, other):
+    def __contains__(self,  other: 'intervalues.BaseInterval | float') -> bool:
         if isinstance(other, int) or isinstance(other, float):
             for key in self.data:
                 if other in key:
@@ -226,7 +252,7 @@ class IntervalSet(AbstractIntervalCollector):
         else:
             raise ValueError(f'Not correct use of "in" for {other}')
 
-    def __getitem__(self, other):
+    def __getitem__(self, other: 'intervalues.BaseInterval | float') -> float:
         if isinstance(other, int) or isinstance(other, float):
             for key in self.data:
                 if other in key:
@@ -243,7 +269,7 @@ class IntervalSet(AbstractIntervalCollector):
         else:
             raise ValueError(f'Not correct use of indexing with {other}')
 
-    def key_compare(self, other):
+    def key_compare(self, other: 'IntervalSet') -> bool:
         keys1, keys2 = sorted(self.data), sorted(other.data)
         while len(keys1) * len(keys2) > 0:
             key1, key2 = keys1.pop(0), keys2.pop(0)
@@ -255,70 +281,52 @@ class IntervalSet(AbstractIntervalCollector):
         return len(keys2) > 0  # shorter before longer - like in BaseInterval
 
     # Implemented to align with BaseInterval ordering, since BaseInterval(0,1) == IntervalCounter((BaseInterval(0,1): 1)
-    def __lt__(self, other):
+    def __lt__(self, other: 'IntervalSet') -> bool:
         other = other.as_set() if not isinstance(other, self.__class__) else other
         return self.key_compare(other)
 
-    def __le__(self, other):
+    def __le__(self, other: 'IntervalSet') -> bool:
         other = other.as_set() if not isinstance(other, self.__class__) else other
         return self == other or self.key_compare(other)
 
-    def __gt__(self, other):
+    def __gt__(self, other: 'IntervalSet') -> bool:
         other = other.as_set() if not isinstance(other, self.__class__) else other
         return other.key_compare(self)
 
-    def __ge__(self, other):
+    def __ge__(self, other: 'IntervalSet') -> bool:
         other = other.as_set() if not isinstance(other, self.__class__) else other
         return self == other or other.key_compare(self)
 
-    def __eq__(self, other):  # Equal if also IntervalSet, with same intervals in it.
+    def __eq__(self, other: object) -> bool:  # Equal if also IntervalSet, with same intervals in it.
         if isinstance(other, type(self)):
             return self.data == other.data
         if isinstance(other, base_interval.BaseInterval) and len(self.data) == 1:
             return other in self.data  # and other.get_length() == self.get_length()
         return False
 
-    def __hash__(self):
+    def __hash__(self) -> int:
         return hash(tuple(self))
 
-    def __iter__(self):
+    def __iter__(self) -> Iterator:
         return iter(self.data)
 
-    def min(self):
+    def min(self) -> float:
         return min(self.data).min()
 
-    def max(self):
+    def max(self) -> float:
         return max(self.data).max()
 
-    def as_meter(self):
+    def as_meter(self) -> 'intervalues.IntervalMeter':
         return intervalues.IntervalMeter(list(iter(self.data)))
 
-    def as_list(self):
+    def as_list(self) -> 'intervalues.IntervalList':
         return intervalues.IntervalList(list(iter(self)))
 
-    def as_counter(self):
+    def as_counter(self) -> 'intervalues.IntervalCounter':
         return intervalues.IntervalCounter(tuple(self))
 
-    def as_pdf(self):
+    def as_set(self) -> 'intervalues.IntervalSet':
+        return self.copy()
+
+    def as_pdf(self) -> 'intervalues.IntervalPdf':
         return intervalues.IntervalPdf(tuple(self))
-
-class IntervalSetFloatTodo(IntervalSet):
-
-    def __call__(self):
-        raise NotImplementedError('__call__ not yet implemented')  # What should it be?
-
-    def draw(self, **kwargs):
-        raise NotImplementedError('To do')  # Draw a value from all intervals - only works if no infinite interval
-
-    def plot(self):
-        raise NotImplementedError('To do')  # Barplot of counts
-
-    def cdf(self, val):
-        raise NotImplementedError('To do')  # Use as cdf: P(X<=val)
-
-    def pdf(self, val):
-        raise NotImplementedError('To do')  # use as pdf: P(X=val) (Integer interval) or Probability of the interval.
-        # Alternative a "Scaled" version that automatically updates the counts to sum/integrate to 1 (not real counts).
-
-    def to_integer_interval(self):
-        raise NotImplementedError('To do')