+++ tests/test_interval_counter.py
@@ -8,7 +8,7 @@ INTERVAL_MANY = [5, 10, 25, 100, 250, 500, 1000, 10000]
 
 def test_init_negative():
     a = BaseInterval(0, 1, -2).as_counter()
-    assert a == EmptyInterval().as_counter()
+    assert a.total_length() == 0
 
     b = IntervalCounter([BaseInterval(0, 1, 3), BaseInterval(2, 3, -1)])
     assert b == IntervalCounter([BaseInterval(0, 1, 3)])
@@ -70,7 +70,7 @@ def test_subtraction_base():
     b = BaseInterval((2, 3))
     c = IntervalCounter([BaseInterval((0, 1)), BaseInterval((2, 3))])
     assert c - b == a
-    assert -b + c == a
+    assert -b + c == c
     c -= b
     assert a == c